Inline sources specific for 222 builds to main source set
