--- src/222/main/resources-nightly/META-INF/platform-experiments.xml
@@ -1,7 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="com.intellij">
-        <experimentalFeature id="org.rust.cargo.evaluate.build.scripts" percentOfUsers="100">
-            <description>Run build scripts while project refresh. It allows collecting information about generated items like `cfg` options, environment variables, etc</description>
-        </experimentalFeature>
-    </extensions>
-</idea-plugin>
