--- src/222/main/resources/META-INF/platform-rust-core.xml
@@ -1,11 +1,0 @@
-<idea-plugin>
-    <extensions defaultExtensionNs="com.intellij">
-        <externalIconProvider key="Cargo"
-                              implementationClass="org.rust.cargo.project.model.impl.CargoExternalSystemIconProvider"/>
-
-        <registryKey key="org.rust.cargo.new.auto.import"
-                     defaultValue="true"
-                     description="Enable new Cargo project model reloading"
-                     restartRequired="true"/>
-    </extensions>
-</idea-plugin>
