+++ src/main/kotlin/org/rust/cargo/project/model/impl/CargoExternalSystemIconProvider.kt
